implement base human
